Extract BaseEntityWithImageManager logic in Trait
